Initial

+++ src/01/axios.ts
@@ -1,17 +1,13 @@
 import axios from 'axios';
+import { createEvent, createStore, forward } from 'effector';
 
 const devUrl = 'https://transparent-staging.herokuapp.com/api';
 const baseURL = process.env.REACT_APP_API_URL || devUrl;
 
 axios.defaults.baseURL = baseURL;
 
-let cancel;
 axios.interceptors.request.use((req) => {
-  if (req.baseURL === 'http://84.201.132.164:8080/api') {
-    delete req.headers.Authorization;
-  } else {
-    req.headers.Authorization = `Bearer ${takeFromLocStor('token')}`;
-  }
+  req.headers.Authorization = `Bearer ${takeFromLocStor('token')}`;
 
   if (req.url && checkUrl('refresh', req.url)) {
     req.data = {
@@ -29,31 +25,40 @@ axios.interceptors.response.use(
     if (url && checkUrl('(login|refresh)', url)) {
       const { token, refreshToken, roles, permissions } = data.successResponse;
 
-      saveToLocStor('token', token);
-      saveToLocStor('refreshToken', refreshToken);
-      saveToLocStor('permissions', permissions);
-      checkUrl('login', url) && saveToLocStor('roles', roles);
+      saveToLocalStorage('token', token);
+      saveToLocalStorage('refreshToken', refreshToken);
+      saveToLocalStorage('permissions', permissions);
+      checkUrl('login', url) && saveToLocalStorage('roles', roles);
     }
 
     if (url && checkUrl('(users/current)', url)) {
       const user = data.successResponse;
-      saveToLocStor('user', user);
+      saveToLocalStorage('user', user);
     }
+
     const res = data.successResponse ?? data ?? {};
+
     return res;
   },
   (error) => {
     const status = error?.response?.status;
     if (status === 401 && !checkUrl('login', error.config.url)) {
       const { config } = error;
+
       return new Promise((resolve, reject) => {
-        axios.post('/auth/refreshToken').then(
-          () => resolve(axios(config)),
-          () => {
-            localStorage.clear();
-            window.location.replace('/login');
-          }
-        );
+        if (!$isRefreshRunning.getState()) {
+          axios.post('/auth/refreshToken').then(
+            () => resolve(axios(config)),
+            () => {
+              localStorage.clear();
+              window.location.replace('/login');
+            }
+          );
+        } else {
+          const subscription = $isRefreshRunning.watch(() => {
+            subscription.unsubscribe();
+          });
+        }
       });
     }
 
@@ -61,19 +66,26 @@ axios.interceptors.response.use(
   }
 );
 
-function saveToLocStor(name: string, data: string) {
+function saveToLocalStorage(name: string, data: string) {
   localStorage.setItem(name, JSON.stringify(data));
 }
 
 function takeFromLocStor(name: string) {
-  const userName = localStorage.getItem(name);
-  if (!userName) return;
-  return JSON.parse(userName);
+  const value = localStorage.getItem(name);
+
+  if (!value) return;
+
+  return JSON.parse(value);
 }
 
 function checkUrl(str: string, url: string) {
   return new RegExp(str, 'gi').test(url);
 }
 
-export { cancel };
+export const $isRefreshRunning = createStore(false);
+
+export const setIsRefreshRunning = createEvent<false>();
+
+$isRefreshRunning.on(setIsRefreshRunning, (_, value) => value);
+
 export default axios;
